Refactor Battle feature to no longer be linked with the object messag…

+++ src/Rhisis.World/Handlers/Battle/RangeAttackHandler.cs
@@ -1,5 +1,6 @@
 ﻿using Rhisis.Game.Abstractions;
 using Rhisis.Game.Abstractions.Entities;
+using Rhisis.Game.Abstractions.Features.Battle;
 using Rhisis.Game.Common;
 using Rhisis.Network;
 using Rhisis.Network.Packets.World;
@@ -46,8 +47,10 @@ namespace Rhisis.World.Handlers.Battle
                 return;
             }
 
-            player.Inventory.DeleteItem(bulletItem, 1);
-            player.Battle.RangeAttack(target, Math.Max(0, packet.Power), ObjectMessageType.OBJMSG_ATK_RANGE1);
+            if(player.Battle.TryRangeAttack(target, Math.Max(0, packet.Power), AttackType.RangeBowAttack))
+            {
+                player.Inventory.DeleteItem(bulletItem, 1);
+            }
         }
     }
 }
